Create README.md
I am really excited to learn more and to connect and talk with people to learn and share experiences and visions. I just recently got in to learning more about crypto in general and the more I get to know the more I think that it is the future and the possibilities and benefits for everyone is endless! 

Looking forward to getting more knowledge.
I hope the person reading this a wonderful day:)